\name{combine_FASTA}
\alias{combine_FASTA}
\title{Combines hmm_fasta objects}
\usage{
  combine_FASTA(protein1, protein2)
}
\arguments{
  \item{protein1}{hmm_fasta object}

  \item{protein2}{hmm_fasta object}
}
\value{
  \item{object}{hmm_fasta object}
}
\description{
  Combines hmm_fasta objects
}
\author{
  Nina Wilkinson
}
\keyword{character}

