\name{hmm.sim}
\alias{equil}
\alias{hmm.sim}
\alias{rdiric}
\title{Simulates a hidden Markov model}
\usage{
  hmm.sim(n, lambda, P, f)
}
\arguments{
  \item{n}{Length of the observed sequence}

  \item{lambda}{The hidden sequence transition matrix}

  \item{P}{An array of transition matrices for observed
  sequence}

  \item{f}{The number of states of observed sequence}
}
\value{
  A list
}
\description{
  Simulates a hidden Markov model
}
\author{
  Nina Wilkinson
}
\keyword{character}

