\name{initialise_checkpoint}
\alias{initialise_checkpoint}
\title{Initialising checkpointing}
\usage{
  initialise_checkpoint(filename, hour)
}
\arguments{
  \item{filename}{name of file for checkpointing}

  \item{hour}{how often checkpointing is required in number
  of iterations}
}
\value{
  \item{cp }{"hmm_checkpoint" object}
}
\description{
  Initialising checkpointing
}
\keyword{character}

