\name{initialise_checkpoint_power}
\alias{initialise_checkpoint_power}
\title{Initialising checkpointing for the power posterior method}
\usage{
  initialise_checkpoint_power(filename)
}
\arguments{
  \item{filename}{filename}
}
\value{
  \item{cp }{"hmm_checkpoint" object}
}
\description{
  Initialising checkpointing for the power posterior method
}
\keyword{character}

