\docType{data}
\name{p53}
\alias{Bnip3}
\alias{CBP}
\alias{EWS}
\alias{FUS}
\alias{mdm2}
\alias{NBR1}
\alias{p53}
\alias{SQSTM1}
\alias{TAF15}
\title{Example FASTA sequences}
\format{Vector of strings}
\source{
  online database NEED MORE DETAILS!!!
}
\description{
  The amino acid sequence of p53, mdm2, CBP, TAF15, FUS,
  EWS, Bnip3, NBR1, SQSTM1. To load a particular sequence,
  use the data function. For example, data(p53)
}
\keyword{datasets}

