\name{read_FASTA}
\alias{read_FASTA}
\alias{reduce_FASTA}
\alias{convert}
\alias{combine_FASTA}
\title{read_FASTA}
\description{get hmm_fasta object}
\usage{read_FASTA(filename = "",level)}
\arguments{
  \item{filename}{location of sequence}
  \item{level}{f}
}
\author{Nina Wilkinson}
\keyword{classif}
