\name{read_FASTA_string}
\alias{convert}
\alias{read_FASTA_string}
\alias{reduce_FASTA}
\title{Makes hmm_fasta object}
\usage{
  read_FASTA_string(name, level)
}
\arguments{
  \item{name}{of vector of strings}

  \item{level}{value of f}
}
\value{
  \item{object}{hmm_fasta object}
}
\description{
  Makes hmm_fasta object
}
\author{
  Nina Wilkinson
}
\keyword{character}

