\name{initialise_prior}
\alias{initialise_prior}
\title{Initialising the prior}
\usage{
  initialise_prior(a, mu, s, r)
}
\arguments{
  \item{a}{prior value for the observed sequence}

  \item{mu}{prior mean of the hidden sequence}

  \item{s}{prior standard deviation for the hidden
  sequence}
}
\value{
  \item{prior }{"prior_parameters" object}
}
\description{
  Initialising the prior
}
\keyword{character}

