\name{read_FASTA}
\alias{read_FASTA}
\title{Makes hmm_fasta object}
\usage{
  read_FASTA(filename, level)
}
\arguments{
  \item{filename}{location of file}

  \item{level}{value of f}
}
\value{
  \item{object}{hmm_fasta object}
}
\description{
  Makes hmm_fasta object
}
\author{
  Nina Wilkinson
}
\keyword{character}

